Using composition to DRY up
