+++ lib/lens/typed.rb
@@ -3,31 +3,32 @@
 module Sorta
   class Lens
     # Extracts data from a given object
-    class Typed < LensBase
+    class Typed
+      include Sorta::Lens::Extractor
+      include Sorta::Lens::Typechecker
+      include Sorta::Lens::Validator
+
+      def self.on(...)
+        new(...)
+      end
+
       def initialize(**kwargs)
         @types = kwargs.values
         validate_types
-        super(*kwargs.keys)
+        @keys = kwargs.keys
+        validate_keys
       end
 
       def call(object)
         @getable = object.respond_to? :[]
         result =
           @keys.zip(@types).each_with_object({}) do |(sym, ty), acc|
-            acc[sym] = extract(sym, object, ty)
+            acc[sym] = extract(sym, object).then do |val|
+              typecheck(val, ty)
+            end
           end
         @getable = nil
         result
-      end
-
-      def extract(sym, object, ty)
-        super(sym, object).then do |val|
-          typecheck(val, ty)
-        end
-      end
-
-      def validate_types
-        validate @types, Class
       end
     end
   end
