+++ lib/lens/untyped.rb
@@ -3,7 +3,19 @@
 module Sorta
   class Lens
     # Extracts data from a given object
-    class Untyped < LensBase
+    class Untyped
+      include Sorta::Lens::Validator
+      include Sorta::Lens::Extractor
+
+      def self.on(...)
+        new(...)
+      end
+
+      def initialize(*keys)
+        @keys = keys
+        validate_keys
+      end
+
       def call(object)
         @getable = object.respond_to? :[]
         result = @keys.each_with_object({}) do |sym, acc|
