+++ lib/sorta-lens.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require_relative "lens/lens"
-require_relative "lens/base"
-require_relative "lens/typed_lens"
+require_relative "lens/components"
+require_relative "lens/typed"
 require_relative "lens/untyped"
 require_relative "version"
 module Sorta
